allow PuppetDB scheme to be changed
In our environment, the PuppetDB server is not publicly visible, and
therefore HTTPS is not enabled with a valid cert. Our prefered way of
accessing the server is through an SSH tunnel, which makes HTTPS
redundant. By allowing the scheme to be changed, we don't have to hack
around the invalid certificate.